requirements.txt
# Microsoft Agent Framework (pin version para estabilidad)
agent-framework-core==1.0.0b260107
openai>=1.0.0

# Servidor HTTP
fastapi>=0.100.0
uvicorn[standard]>=0.22.0

# Utilidades
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0

# Base de datos PostgreSQL
psycopg2-binary>=2.9.0
